file-conv-scripts/pyproject.toml
[tool.poetry]
name = "file-conv-scripts"
version = "0.1.0"
description = ""
authors = ["hermann-web <hermannagossou6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
file-conv-framework = { path = "../file-conv-framework" }

[tool.poetry.group.mimetype.dependencies]
python-magic = "^0.4.27"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"


[tool.poetry.group.dataframe.dependencies]
pandas = "^2.2.1"


[tool.poetry.group.image.dependencies]
pillow = "^10.2.0"


[tool.poetry.group.pdf.dependencies]
pypdf2 = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
